The server with HTML
First run "main.py", at the same time open the HTML file. Paste video-url from TikTok, Bilibili or some other websites, and then click the button. After the page reload, click the sentence which shows in the middle of the screen-it will copy a different url to your clipboard. Paste the new url in the browser and open it, please enjoy the video WITHOUT WATERMARK!!! The video is able to download, too.